// JunctionRelay_Server/junctionrelaywebui/src/types/entities.ts
export interface Sensor {
    id: number;
    originalId: number;
    junctionId?: number | null;
    junctionDeviceLinkId?: number | null;
    junctionCollectorLinkId?: number | null;
    sensorOrder: number;
    mqttServiceId?: number | null;
    mqttTopic?: string | null;
    mqttQoS?: number | null;
    sensorType: string;
    externalId: string;
    deviceName: string;
    name: string;
    componentName: string;
    category: string;
    unit: string;
    value: string;
    decimalPlaces: number;
    sensorTag: string;
    formula?: string | null;
    lastUpdated: string;
    customAttribute1?: string | null;
    customAttribute2?: string | null;
    customAttribute3?: string | null;
    customAttribute4?: string | null;
    customAttribute5?: string | null;
    customAttribute6?: string | null;
    customAttribute7?: string | null;
    customAttribute8?: string | null;
    customAttribute9?: string | null;
    customAttribute10?: string | null;
    isMissing: boolean;
    isStale: boolean;
    isSelected: boolean;
    isVisible: boolean;
    isCustomJunctionSensor: boolean;
    isEventSensor: boolean;
    deviceId?: number | null;
    serviceId?: number | null;
    collectorId?: number | null;
}

export interface Device {
    id: number;
    name: string;
    description: string;
    type: string;
    status: string;
    uniqueIdentifier: string;
    connMode?: string | null;
    comPort?: string | null;
    baudRate?: number | null;
    deviceModel?: string | null;
    deviceManufacturer?: string | null;
    firmwareVersion?: string | null;
    hasCustomFirmware: boolean;
    ignoreUpdates: boolean;
    mcu?: string | null;
    wirelessConnectivity?: string | null;
    flash?: string | null;
    psram?: string | null;
    isConnected: boolean;
    ipAddress?: string | null;
    hasMQTTConfig: boolean;
    httpPort?: number | null;
    webSocketPort?: number | null;
    mqttPort?: number | null;
    hostname?: string | null;
    isGateway: boolean;
    gatewayId?: number | null;
    isJunctionRelayDevice: boolean;
    isCloudDevice: boolean;
    cloudDeviceId?: number | null;
    lastHealthReportAt?: string | null;
    createdAt?: string | null;
    lastHealthAlertSent?: string | null;
    lastHealthReminderSent?: string | null;
    lastUpdated: string;
    pushNotifications: boolean;
    syncMode?: string | null;
    sensors: Sensor[];
    peers: DeviceRecord[];
    pollRate?: number | null;
    sendRate?: number | null;
    sshUsername?: string | null;
    hasSshPassword: boolean;
    externalSshPassword: boolean;
    sshPort?: number | null;
    sshTimeoutMs?: number | null;
    hasSshPrivateKey: boolean;
    externalSshPrivateKey: boolean;
    useSshKeyAuth: boolean;
    sshConnectionRetries?: number | null;
    sshVerifyHostKey: boolean;
    heartbeatProtocol?: string | null;
    heartbeatTarget?: string | null;
    heartbeatExpectedValue?: string | null;
    heartbeatEnabled: boolean;
    heartbeatIntervalMs?: number | null;
    heartbeatGracePeriodMs?: number | null;
    heartbeatMaxRetryAttempts?: number | null;
    useStreamAsHeartbeat: boolean;
    streamHeartbeatThresholdMs?: number | null;
    connectionStatusEnabled: boolean;
    connectionStatusIntervalMs?: number | null;
    lastConnectionStatusCheck?: string | null;
    lastPingAttempt?: string | null;
    lastPinged?: string | null;
    lastPingStatus?: string | null;
    lastPingDurationMs?: number | null;
    consecutivePingFailures?: number | null;
    configLastAppliedAt?: string | null;
    sensorPayloadLastAckAt?: string | null;
    hasOnboardScreen: boolean;
    hasOnboardLED: boolean;
    hasOnboardRGBLED: boolean;
    hasExternalNeopixels: boolean;
    hasExternalMatrix: boolean;
    hasExternalI2CDevices: boolean;
    hasButtons: boolean;
    hasBattery: boolean;
    supportsEthernet: boolean;
    supportsWiFi: boolean;
    supportsBLE: boolean;
    supportsUSB: boolean;
    supportsESPNow: boolean;
    supportsHTTP: boolean;
    supportsMQTT: boolean;
    supportsWebSockets: boolean;
    hasSpeaker: boolean;
    hasMicroSD: boolean;
    isXSD: boolean;
    linkedCollectorId?: number | null;
    screens: DeviceScreen[];
    i2cDevices: DeviceI2CDevice[];
}

export interface DeviceScreen {
    id: number;
    deviceId: number;
    screenKey: string;
    displayName?: string | null;
    screenLayoutId?: number | null;
    template?: Payload | null;
    frameLayoutId?: number | null;
    layoutPath?: string | null;
    frameTemplate?: unknown | null;
    screenType?: string | null;
    supportsConfigPayloads: boolean;
    supportsSensorPayloads: boolean;
    supportsStopPayloads: boolean;
    useKeepAlive?: boolean | null;
}
// ...
export interface Payload {
    id: number;
    displayName?: string | null;
    description?: string | null;
    layoutType: string;
    customLayoutType?: string | null;
    includePrefixConfig: boolean;
    includePrefixSensor: boolean;
    rows?: number | null;
    columns?: number | null;
    jsonLayoutConfig?: string | null;
    isTemplate: boolean;
    isDraft: boolean;
    isPublished: boolean;
    created: string;
    lastModified?: string | null;
    createdBy?: string | null;
    version?: string | null;
    topMargin: number;
    bottomMargin: number;
    leftMargin: number;
    rightMargin: number;
    outerPadding: number;
    innerPadding: number;
    backgroundColor?: string | null;
    borderColor?: string | null;
    borderVisible?: boolean | null;
    borderThickness?: number | null;
    roundedCorners?: boolean | null;
    borderRadiusSize?: number | null;
    opacityPercentage?: number | null;
    gradientDirection?: string | null;
    gradientEndColor?: string | null;
    chartOutlineVisible?: boolean | null;
    showLegend?: boolean | null;
    positionLegendInside?: boolean | null;
    showXAxisLabels?: boolean | null;
    showYAxisLabels?: boolean | null;
    gridDensity?: number | null;
    historyPointsToShow?: number | null;
    chartScrollSpeed?: number | null;
    showUnits?: boolean | null;
    textColor?: string | null;
    textSize?: string | null;
    labelSize?: string | null;
    valueSize?: string | null;
    decimalPlaces?: number | null;
    justifyContent?: string | null;
    alignItems?: string | null;
    textAlignment?: string | null;
    animationType?: string | null;
    animationDuration?: number | null;
    showPreview: boolean;
    previewHeight?: number | null;
    previewWidth?: number | null;
    previewSensors?: number | null;
    isResponsive: boolean;
    mobileLayoutBehavior?: string | null;
    themeId?: number | null;
    inheritThemeStyles: boolean;
    allowInteraction: boolean;
    onClickBehavior?: string | null;
    navigationTarget?: string | null;
    dataRefreshIntervalSeconds?: number | null;
    cacheData: boolean;
    dataFilterCriteria?: string | null;
    backgroundImageUrl?: string | null;
    backgroundImageId?: string | null;
    imageFit?: string | null;
    lazyLoad: boolean;
    renderPriority?: number | null;
    enableScrollbars: boolean;
    minWidth?: number | null;
    maxWidth?: number | null;
    minHeight?: number | null;
    maxHeight?: number | null;
    fieldsToSend?: string | null;
}
// ...
export interface DeviceI2CDevice {
    id: number;
    deviceId: number;
    i2CAddress: string;
    deviceType: string;
    communicationProtocol: string;
    isEnabled: boolean;
    dateAdded: string;
    endpoints: DeviceI2CDeviceEndpoint[];
}
// ...
export interface DeviceRecord {
    id: number;
    name: string;
    description: string;
    type: string;
    status: string;
    uniqueIdentifier: string;
    connMode?: string | null;
    comPort?: string | null;
    baudRate?: number | null;
    deviceModel?: string | null;
    deviceManufacturer?: string | null;
    firmwareVersion?: string | null;
    hasCustomFirmware: boolean;
    ignoreUpdates: boolean;
    mcu?: string | null;
    wirelessConnectivity?: string | null;
    flash?: string | null;
    psram?: string | null;
    isConnected: boolean;
    ipAddress?: string | null;
    hasMQTTConfig: boolean;
    httpPort?: number | null;
    webSocketPort?: number | null;
    mqttPort?: number | null;
    hostname?: string | null;
    isGateway: boolean;
    gatewayId?: number | null;
    isJunctionRelayDevice: boolean;
    isCloudDevice: boolean;
    cloudDeviceId?: number | null;
    lastEncryptedSensorData?: string | null;
    lastHealthReportAt?: string | null;
    createdAt?: string | null;
    lastHealthAlertSent?: string | null;
    lastHealthReminderSent?: string | null;
    pushNotifications: boolean;
    syncMode?: string | null;
    lastUpdated: string;
    sensors: Sensor[];
    peers: DeviceRecord[];
    pollRate?: number | null;
    sendRate?: number | null;
    sshUsername?: string | null;
    sshPassword?: string | null;
    externalSshPassword: boolean;
    sshPort?: number | null;
    sshTimeoutMs?: number | null;
    sshPrivateKey?: string | null;
    externalSshPrivateKey: boolean;
    useSshKeyAuth: boolean;
    sshConnectionRetries?: number | null;
    sshVerifyHostKey: boolean;
    heartbeatProtocol?: string | null;
    heartbeatTarget?: string | null;
    heartbeatExpectedValue?: string | null;
    heartbeatEnabled: boolean;
    heartbeatIntervalMs?: number | null;
    heartbeatGracePeriodMs?: number | null;
    heartbeatMaxRetryAttempts?: number | null;
    useStreamAsHeartbeat: boolean;
    streamHeartbeatThresholdMs?: number | null;
    connectionStatusEnabled: boolean;
    connectionStatusIntervalMs?: number | null;
    lastConnectionStatusCheck?: string | null;
    lastPingAttempt?: string | null;
    lastPinged?: string | null;
    lastPingStatus?: string | null;
    lastPingDurationMs?: number | null;
    consecutivePingFailures?: number | null;
    configLastAppliedAt?: string | null;
    sensorPayloadLastAckAt?: string | null;
    hasOnboardScreen: boolean;
    hasOnboardLED: boolean;
    hasOnboardRGBLED: boolean;
    hasExternalNeopixels: boolean;
    hasExternalMatrix: boolean;
    hasExternalI2CDevices: boolean;
    hasButtons: boolean;
    hasBattery: boolean;
    supportsEthernet: boolean;
    supportsWiFi: boolean;
    supportsBLE: boolean;
    supportsUSB: boolean;
    supportsESPNow: boolean;
    supportsHTTP: boolean;
    supportsMQTT: boolean;
    supportsWebSockets: boolean;
    hasSpeaker: boolean;
    hasMicroSD: boolean;
    isXSD: boolean;
    linkedCollectorId?: number | null;
    screens: DeviceScreen[];
    i2cDevices: DeviceI2CDevice[];
}
// ...
export interface DeviceI2CDeviceEndpoint {
    id: number;
    i2CDeviceId: number;
    endpointType: string;
    address: string;
    qoS: number;
    notes?: string | null;
}